feat: implemented function that remembering last place's position when app restarted - Focus same place as the last place before termination of application - remember position using sharedPreference

+++ app/src/main/java/campus/tech/kakao/map/SearchActivity.kt
@@ -120,8 +120,8 @@ class SearchActivity : AppCompatActivity() {
 
     private fun goBackToMap() {
         val searchToMapIntent = Intent(this, MapActivity::class.java)
-        searchToMapIntent.putExtra("mapX",mapX.toDouble())
-        searchToMapIntent.putExtra("mapY",mapY.toDouble())
+        searchToMapIntent.putExtra("mapX",mapX)
+        searchToMapIntent.putExtra("mapY",mapY)
         searchToMapIntent.putExtra("name",name)
         Log.d("goBackToMap", "goBackToMap: $mapX, $mapY")
         startActivity(searchToMapIntent)
